+++ CHANGELOG.md
@@ -85,6 +85,7 @@ This release contains several breaking changes to 1.6 due to heavy internal refa
 - Windows installers are finally signed.
 - Switched back to the `package.json` configuration for electron-builder, because, well, Electron.
 - Fix a lot of unused and weirdly coded code.
+- Added a flag to determine if the Zettlr version has changed, indicating an update (or downgrade). Based on this, the FSAL will clear its cache on boot.
 
 # 1.6.0